+++ rskj-core/src/main/java/org/ethereum/core/Repository.java
@@ -146,13 +146,13 @@ public interface Repository extends RepositorySnapshot {
         addBalance(toAddr, value);
     }
 
-    RentedNode getRentedNode(TrackedNode trackedNode);
+//    RentedNode getRentedNode(TrackedNode trackedNode);
 
-    void updateRents(Set<RentedNode> rentedNodes, long executionBlockTimestamp);
+//    void updateRents(Set<RentedNode> rentedNodes, long executionBlockTimestamp);
 
     // todo(fedejinich) this methods will only be included in MutableRepositoryTracked
 
-    Set<TrackedNode> getStorageRentNodes(String transactionHash);
+//    Set<TrackedNode> getStorageRentNodes(String transactionHash);
 
-    List<TrackedNode> getRollBackNodes(String transactionHash);
+//    List<TrackedNode> getRollBackNodes(String transactionHash);
 }
