+++ rskj-core/src/main/java/org/ethereum/db/MutableRepository.java
@@ -60,55 +60,26 @@ public class MutableRepository implements Repository {
     protected final TrieKeyMapper trieKeyMapper;
     protected final MutableTrie mutableTrie;
 
-    // todo(fedejinich) ALL THIS MEMBERS MIGHT BE MOVED TO MutableRepositoryTracked
-    // enables node tracking feature
-    private final boolean enableTracking;
-    // a set to track all the used trie-value-containing nodes in this repository (and its children repositories)
-    protected final Set<TrackedNode> trackedNodes;
-    // a list of nodes tracked nodes that were rolled back (due to revert or OOG)
-    protected final List<TrackedNode> rollbackNodes;
-    // parent repository to commit tracked nodes
-    protected final MutableRepository parentRepository;
-    // this contains the hash of the ongoing tracked transaction
-    protected String trackedTransactionHash = "NO_TRANSACTION_HASH";
-
     // default constructor
-    protected MutableRepository(MutableTrie mutableTrie, MutableRepository parentRepository,
-                                Set<TrackedNode> trackedNodes, List<TrackedNode> rollbackNodes,
-                                boolean enableTracking) {
+    public MutableRepository(MutableTrie mutableTrie) {
         this.trieKeyMapper = new TrieKeyMapper();
         this.mutableTrie = mutableTrie;
-        this.parentRepository = parentRepository;
-        this.trackedNodes = trackedNodes;
-        this.rollbackNodes = rollbackNodes;
-        this.enableTracking = enableTracking;
     }
 
-    // creates a new repository (tracking disabled)
-    public MutableRepository(MutableTrie mutableTrie) {
-        this(mutableTrie,  null, Collections.emptySet(), Collections.emptyList(), false);
-    }
+//    // creates a new repository (tracking disabled)
+//    public MutableRepository(MutableTrie mutableTrie) {
+//        this(mutableTrie,  null, Collections.emptySet(), Collections.emptyList(), false);
+//    }
 
     // another way to create a repository (tracking disabled)
     public MutableRepository(TrieStore trieStore, Trie trie) {
-        this(new MutableTrieImpl(trieStore, trie), null, new HashSet<>(), new ArrayList<>(), false);
+        this(new MutableTrieImpl(trieStore, trie));//, null, new HashSet<>(), new ArrayList<>(), false);
     }
 
     // another way to create a repository (tracking disabled),
     // by using this way we avoid instantiating Trie instances from undesired class
     public MutableRepository(TrieStore trieStore) {
         this(trieStore, new Trie(trieStore));
-    }
-
-    // useful for key tracking, it connects between repositories. the root Repository will contain a null parentRepository
-    protected MutableRepository(MutableTrie mutableTrie, MutableRepository parentRepository, boolean enableTracking) {
-        this(mutableTrie, parentRepository, new HashSet<>(), new ArrayList<>(), enableTracking);
-    }
-
-    // creates a tracked repository, all the child repositories (created with startTracking()) will also be tracked.
-    // this should be only called from RepositoryLocator.trackedRepository
-    public static MutableRepository trackedRepository(MutableTrie mutableTrieCache) {
-        return new MutableRepository(mutableTrieCache, null, true);
     }
 
     @Override
@@ -366,9 +337,7 @@ public class MutableRepository implements Repository {
     // To start tracking, a new repository is created, with a MutableTrieCache in the middle
     @Override
     public synchronized Repository startTracking() {
-        MutableRepository mutableRepository = new MutableRepository(new MutableTrieCache(this.mutableTrie), this, this.enableTracking);
-        mutableRepository.setTrackedTransactionHash(trackedTransactionHash); // todo(fedejinich) this will be moved to MutableRepositoryTracked
-        return mutableRepository;
+        return new MutableRepository(new MutableTrieCache(this.mutableTrie));//, this, this.enableTracking);
     }
 
     @Override
@@ -379,22 +348,11 @@ public class MutableRepository implements Repository {
     @Override
     public synchronized void commit() {
         this.mutableTrie.commit();
-
-        if(this.parentRepository != null) {
-            this.parentRepository.mergeTrackedNodes(this.trackedNodes);
-            this.parentRepository.addRollbackNodes(this.rollbackNodes);
-        }
     }
 
     @Override
     public synchronized void rollback() {
         this.mutableTrie.rollback();
-
-        if(parentRepository != null) {
-            this.parentRepository.addRollbackNodes(this.trackedNodes);
-            this.trackedNodes.clear();
-            this.rollbackNodes.clear();
-        }
     }
 
     @Override
@@ -438,165 +396,28 @@ public class MutableRepository implements Repository {
         return storageRootNode.getHash().getBytes();
     }
 
-    // todo(fedejinich) all the content below might be extracted to MutableRepositoryTracked
-
-    @VisibleForTesting
-    public Set<TrackedNode> getTrackedNodes() {
-        return this.trackedNodes;
-    }
-
-    public List<TrackedNode> getRollBackNodes(String transactionHash) {
-        return this.rollbackNodes.stream()
-                .filter(trackedNode -> trackedNode.useForStorageRent(transactionHash))
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public RentedNode getRentedNode(TrackedNode trackedNode) {
-        byte[] key = trackedNode.getKey().getData();
-
-        // if we reach here, it will always get timestamp/valueLength from an existing key
-
-        Long nodeSize = Long.valueOf(this.mutableTrie.getValueLength(key).intValue());
-        Optional<Long> rentTimestamp = this.mutableTrie.getRentTimestamp(key);
-        long lastRentPaidTimestamp = rentTimestamp.isPresent() ? rentTimestamp.get() : NO_RENT_TIMESTAMP;
-
-        RentedNode rentedNode = new RentedNode(trackedNode, nodeSize, lastRentPaidTimestamp);
-
-        return rentedNode;
-    }
-
-    @Override
-    public void updateRents(Set<RentedNode> rentedNodes, long executionBlockTimestamp) {
-        rentedNodes.forEach(node -> {
-            long updatedRentTimestamp = node.getUpdatedRentTimestamp(executionBlockTimestamp);
-
-            this.mutableTrie.putRentTimestamp(node.getKey().getData(), updatedRentTimestamp);
-        });
-    }
-
-    public void setTrackedTransactionHash(String trackedTransactionHash) {
-        this.trackedTransactionHash = trackedTransactionHash;
-    }
-
-    // Internal methods contains node tracking
-
     protected void internalPut(byte[] key, byte[] value) {
         mutableTrie.put(key, value);
-        if(value == null) {
-            trackNodeDeleteOperation(key);
-        } else {
-            trackNodeWriteOperation(key);
-        }
     }
 
     protected void internalDeleteRecursive(byte[] key) {
         // todo(fedejinich) what happens for non existing keys? should track with false result?
         mutableTrie.deleteRecursive(key);
-        trackNodeDeleteOperation(key);
     }
 
     protected byte[] internalGet(byte[] key, boolean readsContractCode) {
-        byte[] value = mutableTrie.get(key);
-        boolean isSuccessful = value != null;
-
-        if(readsContractCode) {
-            trackNodeReadContractOperation(key, isSuccessful);
-        } else {
-            // todo(fedejinich) should track get() success with a bool (value != null)
-            trackNodeReadOperation(key, isSuccessful);
-        }
-
-        return value;
+        return mutableTrie.get(key);
     }
 
     protected Optional<Keccak256> internalGetValueHash(byte[] key) {
-        Optional<Keccak256> valueHash = mutableTrie.getValueHash(key);
-
-        trackNodeReadOperation(key, valueHash.isPresent());
-
-        return valueHash;
+        return mutableTrie.getValueHash(key);
     }
 
     protected Uint24 internalGetValueLength(byte[] key) {
-        Uint24 valueLength = mutableTrie.getValueLength(key);
-
-        trackNodeReadOperation(key, valueLength != Uint24.ZERO);
-
-        return valueLength;
+        return mutableTrie.getValueLength(key);
     }
 
     protected Iterator<DataWord> internalGetStorageKeys(RskAddress addr) {
-        Iterator<DataWord> storageKeys = mutableTrie.getStorageKeys(addr);
-
-        // todo(fedejinich) how should I track the right key/s?
-        boolean result = !storageKeys.equals(Collections.emptyIterator());
-        byte[] accountStoragePrefixKey = trieKeyMapper.getAccountStoragePrefixKey(addr);
-
-        trackNodeReadOperation(accountStoragePrefixKey, result);
-
-        return storageKeys;
-    }
-
-    protected void trackNodeWriteOperation(byte[] key) {
-        trackNode(key, WRITE_OPERATION, true);
-    }
-
-    protected void trackNodeDeleteOperation(byte[] key) {
-        trackNode(key, DELETE_OPERATION, true);
-    }
-
-    protected void trackNodeReadOperation(byte[] key, boolean result) {
-        trackNode(key, READ_OPERATION, result);
-    }
-
-    protected void trackNodeReadContractOperation(byte[] key, boolean result) {
-        trackNode(key, READ_CONTRACT_CODE_OPERATION, result);
-    }
-
-    protected void trackNode(byte[] key, OperationType operationType, boolean isSuccessful) {
-        if(this.enableTracking) {
-            TrackedNode trackedNode = new TrackedNode(
-                new ByteArrayWrapper(key),
-                operationType,
-                this.trackedTransactionHash,
-                isSuccessful
-            );
-            this.trackedNodes.add(trackedNode);
-        }
-    }
-
-    private void mergeTrackedNodes(Set<TrackedNode> trackedNodes) {
-        this.trackedNodes.addAll(trackedNodes);
-    }
-
-    private void addRollbackNodes(Collection<TrackedNode> trackedNodes) {
-        this.rollbackNodes.addAll(trackedNodes);
-    }
-
-    @Override
-    public Set<TrackedNode> getStorageRentNodes(String transactionHash) {
-        Map<ByteArrayWrapper, TrackedNode> storageRentNodes = new HashMap<>();
-        this.trackedNodes.stream()
-                .filter(trackedNode -> trackedNode.useForStorageRent(transactionHash))
-                .forEach(trackedNode -> {
-                    ByteArrayWrapper key = new ByteArrayWrapper(trackedNode.getKey().getData());
-                    TrackedNode containedNode = storageRentNodes.get(key);
-
-                    boolean isContainedNode = containedNode != null;
-                    if(isContainedNode) {
-                        long notRelevant = -1;
-                        RentedNode nodeToBeReplaced = new RentedNode(containedNode, notRelevant, notRelevant);
-                        RentedNode newNode = new RentedNode(trackedNode, notRelevant, notRelevant);
-                        if(nodeToBeReplaced.shouldBeReplaced(newNode)) {
-                            // we pass the TrackedNode instance because we don't need a populated RentedNode yet
-                            storageRentNodes.put(key, trackedNode);
-                        }
-                    } else {
-                        storageRentNodes.put(key, trackedNode);
-                    }
-                });
-
-        return new HashSet<>(storageRentNodes.values());
+        return mutableTrie.getStorageKeys(addr);
     }
 }
